♻️ (dashboard): Move Overview KPI derivation to a sibling module
Extract `deriveOverviewKpis` + `compareBySeverity` + `OverviewKpis` into a plain
`OverviewPage.kpis.ts` module (matching the repo's fleetTypes/fleetFilters
pattern). Keeps `OverviewPage.tsx` a components-only module so it satisfies
react-refresh `only-export-components`, and makes the pure derivation directly
unit-testable. No behaviour change.
